Added the current files.
v2.0.ps is the video downloader at its most complete state. The script can download using an M3U8 link, or for multiple videos it can import a series of file from the videos.csv file if the command "csv" is entered.
The File-Getter.ps is a super short offshoot from a few version ago which will download a single file with a loading bar.